[Theme] Moved Loader namespace to Configuration\*
