+++ src/Sylius/Bundle/ThemeBundle/spec/Configuration/Processor/SymfonyConfigurationProcessorSpec.php
@@ -9,11 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Sylius\Bundle\ThemeBundle\Loader;
+namespace spec\Sylius\Bundle\ThemeBundle\Configuration\Processor;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Bundle\ThemeBundle\Loader\ConfigurationProcessorInterface;
-use Sylius\Bundle\ThemeBundle\Loader\SymfonyConfigurationProcessor;
+use Sylius\Bundle\ThemeBundle\Configuration\Processor\ConfigurationProcessorInterface;
+use Sylius\Bundle\ThemeBundle\Configuration\Processor\SymfonyConfigurationProcessor;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 use Symfony\Component\Config\Definition\Processor;
 
@@ -31,7 +31,7 @@ class SymfonyConfigurationProcessorSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('Sylius\Bundle\ThemeBundle\Loader\SymfonyConfigurationProcessor');
+        $this->shouldHaveType('Sylius\Bundle\ThemeBundle\Configuration\Processor\SymfonyConfigurationProcessor');
     }
 
     function it_implements_configuration_processor_interface()
